fix warnings

+++ Utils/GLApp.h
@@ -100,7 +100,7 @@ public:
   Mat4 getDrawTransform() const;
 
   void resetPointTexture(uint32_t resolution=64);
-  void setPointTexture(const std::vector<uint8_t>& shape, uint32_t x, uint32_t y, uint32_t components);
+  void setPointTexture(const std::vector<uint8_t>& shape, uint32_t x, uint32_t y, uint8_t components);
   void setPointTexture(const Image& shape);
   void setPointHighlightTexture(const Image& shape);
   void resetPointHighlightTexture();
